Change score function

+++ Kernel/PlotData.h
@@ -33,7 +33,7 @@ public:
   };
 
   CPointData max0() const;
-  double score() const;
+  int score() const;
 
   void fillPlots(const CContainer& Samples);
